requirements.txt
pandas
python-dotenv
openai
nbformat
nbconvert
numpy
boto3